Refactor existing monsters endpoint; add logger
As a part of the preparation to begin the
implementation of the authorization and
authentication endpoint, I went ahead and
refactored and restructured the monsters endpoint
to facilitate a generally better development
flow for endpoints.

I also added a new logger to give me timestamped
and colour-indicated debug logs throughout the
runtime of the API. I figure this will be useful
during debugging.